qcacld-3.0: Avoid crash when psoc trans is rejected
Crash is seen when the mode change and shutdown happens in
parallel. Mode change is the driver trans and shutdown is a
psoc trans. If mode change is going on and shutdown req comes
then shutddown trans is rejected. Currently driver crashes
whenever psoc trans is rejected.

Avoid BUG_ON when the psoc trans is rejected.

CRs-Fixed: 2732103
Change-Id: Iad1ab4f188702ed581e3b41c1cdef1c495015048